move config-updating code to new module
